Updated test case for birthday field
In the current implementation, it's not reflected that the month is 0-based in JavaScript dates (see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#Parameters). This is not really covered by the test.

Furthermore, there is a problem with daylight savings if the new Date() object is just instanciated with 

    new Date(1678, 4, 24)

It would make sense to use moment.js to make it more sold, and avoid having casting problems.